+++ client/web/src/enterprise/insights/core/backend/gql-backend/gql/GetInsights.ts
@@ -40,6 +40,7 @@ const INSIGHT_VIEW_SERIES_FRAGMENT = gql`
 export const INSIGHT_VIEW_FRAGMENT = gql`
     fragment InsightViewNode on InsightView {
         id
+        isFrozen
         appliedFilters {
             includeRepoRegex
             excludeRepoRegex