+++ client/web/src/integration/insights/fixtures/insights-metadata.ts
@@ -8,6 +8,7 @@ export const createJITMigrationToGQLInsightMetadataFixture = (options: InsightOp
     __typename: 'InsightView',
     id: '001',
     dashboardReferenceCount: 0,
+    isFrozen: false,
     appliedFilters: {
         __typename: 'InsightViewFilters',
         includeRepoRegex: '',
@@ -71,6 +72,7 @@ export const STORYBOOK_GROWTH_INSIGHT_METADATA_FIXTURE: InsightViewNode = {
     __typename: 'InsightView',
     id: '002',
     dashboardReferenceCount: 0,
+    isFrozen: false,
     appliedFilters: {
         __typename: 'InsightViewFilters',
         includeRepoRegex: '',
@@ -112,6 +114,7 @@ export const SOURCEGRAPH_LANG_STATS_INSIGHT_METADATA_FIXTURE: InsightViewNode = 
     __typename: 'InsightView',
     id: '003',
     dashboardReferenceCount: 0,
+    isFrozen: false,
     appliedFilters: {
         __typename: 'InsightViewFilters',
         includeRepoRegex: '',