+++ client/web/src/views/components/view/content/chart-view-content/ChartViewContent.tsx
@@ -38,13 +38,20 @@ export interface ChartViewContentProps {
      * the chart datum (pie arc, line point, bar category)
      */
     onDatumLinkClick?: () => void
+    locked?: boolean
 }
 
 /**
  * Display chart content with different type of charts (line, bar, pie)
  */
 export const ChartViewContent: FunctionComponent<ChartViewContentProps> = props => {
-    const { content, className = '', layout = ChartViewContentLayout.ByParentSize, onDatumLinkClick = noop } = props
+    const {
+        content,
+        className = '',
+        layout = ChartViewContentLayout.ByParentSize,
+        onDatumLinkClick = noop,
+        locked = false,
+    } = props
 
     // Click link-zone handler for line chart only. Catch click around point and redirect user by
     // link which we've got from the nearest datum point to user cursor position. This allows user
@@ -80,19 +87,32 @@ export const ChartViewContent: FunctionComponent<ChartViewContentProps> = props 
                                 width={width}
                                 height={height}
                                 hasChartParentFixedSize={isResponsive}
+                                locked={locked}
                             />
                         )
                     }
 
                     if (content.chart === 'bar') {
                         return (
-                            <BarChart {...content} width={width} height={height} onDatumLinkClick={onDatumLinkClick} />
+                            <BarChart
+                                {...content}
+                                width={width}
+                                height={height}
+                                onDatumLinkClick={onDatumLinkClick}
+                                locked={locked}
+                            />
                         )
                     }
 
                     if (content.chart === 'pie') {
                         return (
-                            <PieChart {...content} width={width} height={height} onDatumLinkClick={onDatumLinkClick} />
+                            <PieChart
+                                {...content}
+                                width={width}
+                                height={height}
+                                onDatumLinkClick={onDatumLinkClick}
+                                locked={locked}
+                            />
                         )
                     }