+++ client/web/src/views/components/view/content/chart-view-content/charts/bar/BarChart.tsx
@@ -11,6 +11,7 @@ import classNames from 'classnames'
 import { range } from 'lodash'
 import { BarChartContent } from 'sourcegraph'
 
+import { LockedChart } from '../locked/LockedChart'
 import { MaybeLink } from '../MaybeLink'
 
 import styles from './BarChart.module.scss'
@@ -36,6 +37,7 @@ interface BarChartProps<Datum extends object> extends Omit<BarChartContent<Datum
     height: number
     /** Callback calls every time when a bar-link on the chart was clicked */
     onDatumLinkClick: (event: React.MouseEvent) => void
+    locked?: boolean
 }
 
 /**
@@ -49,6 +51,7 @@ export function BarChart<Datum extends object>(props: BarChartProps<Datum>): Rea
         series,
         onDatumLinkClick,
         xAxis: { dataKey: xDataKey },
+        locked = false,
     } = props
 
     // Respect only first element of data series
@@ -98,6 +101,10 @@ export function BarChart<Datum extends object>(props: BarChartProps<Datum>): Rea
         tooltipTimeout = window.setTimeout(() => {
             hideTooltip()
         }, 300)
+    }
+
+    if (locked) {
+        return <LockedChart />
     }
 
     return (